# openapi.yaml
openapi: 3.1.0
info:
  title: Вместе API
  version: 1.0.0
  description: >-
    HTTP-контракт маршрутов приложения. Идентификаторы userId, organizerId и
    другие actor ID, переданные клиентом, сами по себе не обеспечивают
    аутентификацию. Подписанные MAX initData проверяются только на операциях
    сохранения, чтения и удаления туристических планов.
servers:
  - url: https://maxserver-iwrawww.amvera.io
paths:
...
  /api/events:
    get:
      summary: Получение ленты событий
      responses:
        "200":
          description: Список событий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      summary: Создание события
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
      responses:
        "201":
          description: Событие создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Некорректные данные события
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    parameters:
      - name: city
        in: query
        schema:
          type: string
      - name: category
        in: query
        schema:
          type: string
      - name: price
        in: query
        schema:
          type: string
      - name: viewerId
        in: query
        schema:
          type: string
  /api/events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Получение события
      responses:
        "200":
          description: Событие найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Изменение события
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Событие обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Изменение запрещено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Событие завершено или данные некорректны
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Удаление события организатором
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Событие удалено
        "403":
          description: Удаление запрещено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Событие завершено и не может быть удалено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/events/{eventId}/join:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: Запись пользователя на событие
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventParticipation"
      responses:
        "200":
          description: Пользователь записан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResult"
        "400":
          description: Запись невозможна
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/events/{eventId}/leave:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: Отмена участия в событии
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventParticipation"
      responses:
        "200":
          description: Участие отменено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResult"
        "400":
          description: Пользователь не участвует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/events/joined:
    get:
      summary: Получение текущих участий пользователя
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Идентификаторы событий, в которых пользователь участвует
          content:
            application/json:
              schema:
                type: object
                required: [eventIds]
                properties:
                  eventIds:
                    type: array
                    items:
                      type: integer
  /api/events/participated:
    get:
      summary: Получение истории участий пользователя
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Идентификаторы событий, в которых пользователь участвовал
          content:
            application/json:
              schema:
                type: object
                required: [eventIds]
  /api/events/{eventId}/participants:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Получение списка участников события
      responses:
        "200":
          description: Участники и организатор
          content:
            application/json:
              schema:
                type: object
                required: [participants]
                properties:
                  participants:
                    type: array
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/events/{eventId}/reviews:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Получение отзывов о событии
      responses:
        "200":
          description: Список отзывов; для будущего события пуст
          content:
            application/json:
              schema:
                type: array
        "404":
          description: Событие не найдено
    post:
      summary: Добавление отзыва о завершённом событии
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, rating, text]
      responses:
        "201":
          description: Отзыв добавлен
        "403":
          description: Событие ещё не завершено или отзыв запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Некорректный отзыв или пользователь уже оставил отзыв
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    maxInitData:
      type: apiKey
      in: header
      name: Authorization
      description: "MAX initData передаётся как Authorization: tma <initData>."
  schemas:
...
    Event:
      type: object
      required: [id, title, category, date, price, participants]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        date:
          type: string
        price:
          type: string
        participants:
          type: integer
        organizerId:
          type: string
        updatedAt:
          type: string
          format: date-time
      additionalProperties: true
    EventCreate:
      type: object
      required: [title, organizerId]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        date:
          type: string
        duration:
          type: string
        format:
          type: string
          enum: [Онлайн, Офлайн]
        price:
          type: string
          enum: [Бесплатно, Платно, Пушкинская карта]
        city:
          type: string
        maxParticipants:
          type: integer
        organizerId:
          type: string
      additionalProperties: true
    EventUpdate:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        title:
          type: string
        description:
          type: string
        address:
          type: string
        format:
          type: string
        duration:
          type: string
      additionalProperties: true
    EventParticipation:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
      additionalProperties: true
    ParticipationResult:
      type: object
      required: [success, participants]
      properties:
        success:
          type: boolean
        participants:
          type: integer
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string